[java] remove deprecated Firefox capabilities and consolidate tests
