+++ java/test/org/openqa/selenium/firefox/BUILD.bazel
@@ -4,7 +4,6 @@ load("//java:defs.bzl", "JUNIT5_DEPS", "java_selenium_test_suite", "java_test_su
 
 LARGE_TESTS = [
     "FirefoxDriverTest.java",
-    "MarionetteTest.java",
     "TakesFullPageScreenshotTest.java",
 ]
